[Done] DiKantin WEB

+++ database/migrations/2014_10_05_030646_create_role_has_model.php
@@ -11,7 +11,7 @@ return new class extends Migration {
     public function up(): void
     {
         Schema::create('model_has_roles', function (Blueprint $table) {
-            $table->integer('role_id');
+            $table->unsignedBigInteger('role_id');
             $table->string('model_type', 100);
             $table->integer('model_id');
             $table->foreign('role_id')->references('id')->on('roles')->cascadeOnDelete()->cascadeOnUpdate();
